+++ src/AttributedDI.SourceGenerator/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using static AttributedDI.SourceGenerator.Utils.IncrementalValuesProviderHelper;
