+++ src/AttributedDI.SourceGenerator/ServiceRegistrationGenerator.cs
@@ -1,5 +1,6 @@
+using AttributedDI.SourceGenerator.InterfacesGeneration;
+using AttributedDI.SourceGenerator.ServiceModulesGeneration;
 using Microsoft.CodeAnalysis;
-using System.Collections.Immutable;
 using System.Linq;
 
 namespace AttributedDI.SourceGenerator;
